Removes conflicts and shifts to tests_views.py
